Add files via upload
Daily expense app with user Authentication

+++ Expense App--Backened/controllers/expense.js
@@ -5,7 +5,7 @@ const addexpense = (req, res) => {
     if(expenseamount == undefined || expenseamount.length === 0){
         return res.status(400).json({success:false, message: 'parameter is missing'})
     }
-    Expense.create({expenseamount, description, category}).then(expense => {
+    Expense.create({expenseamount, description, category, userId:req.user.id}).then(expense => {
         return res.status(201).json({expense, success:true});
     }).catch(err => {
         return res.status(500).json({success:false, error:err});
@@ -13,7 +13,7 @@ const addexpense = (req, res) => {
 }
 
 const getexpenses = (req, res) => {
-      Expense.findAll().then(expenses => {
+      Expense.findAll( { where: {userId: req.user.id}}).then(expenses => {
         return res.status(200).json({expenses, success:true});
       })
     .catch(err =>{
@@ -27,7 +27,7 @@ const deleteexpense = (req, res) => {
     if(expenseid == undefined || expenseid.length == 0){
         res.status(400).json({success:false,})
     }
-    Expense.destroy({where: { id: expenseid}}).then(() => {
+    Expense.destroy({where: { id: expenseid, userId:req.user.id}}).then(() => {
         return res.status(200).json({success:true, message: "Deleted Successfully"});
     }).catch(err => {
         console.log(err);
